Feature/194 bug fix (#195)
* 병합 실수로 인한 코드 누락 수정,
검증안된 userInfo 객체 검증후 사용하도록 수정,companyBankName 은행명 속성 추가#194

* 로그인 완료 이후 json 데이터 출력 오류 수정#194

+++ src/main/java/com/danakga/webservice/user/repository/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<UserInfo,Integer> {
+public interface UserRepository extends JpaRepository<UserInfo,Long> {
     Optional<UserInfo> findByUserid(String userid);
 
     Optional<UserInfo> findByEmail(String email);
